Local_leyes_RD/leyesApp/requirements.txt
requests
pandas
beautifulsoup4
boto3
html5lib
supabase
pdfplumber
openai
docx
unidecode
python-docx
psycopg2-binary